Has moved files from another folder
