+++ app/src/main/java/org/secfirst/umbrella/rss/adapters/ArticleSimpleAdapter.java
@@ -1,4 +1,4 @@
-package org.secfirst.umbrella.rss.feed.adapters;
+package org.secfirst.umbrella.rss.adapters;
 
 import android.content.Context;
 import android.content.Intent;
@@ -17,7 +17,7 @@ import com.einmalfel.earl.Item;
 
 import org.secfirst.umbrella.R;
 import org.secfirst.umbrella.WebViewActivity;
-import org.secfirst.umbrella.rss.feed.CustomFeed;
+import org.secfirst.umbrella.rss.entities.CustomFeed;
 import org.secfirst.umbrella.util.UmbrellaUtil;
 
 import java.util.ArrayList;
