+++ app/src/main/res/layout/activity_article.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:background="@color/light_grey"
-    tools:context="org.secfirst.umbrella.rss.feed.ArticleActivity">
+    tools:context="org.secfirst.umbrella.rss.views.ArticleActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
