+++ app/src/main/res/menu/custom_feed_list.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="org.secfirst.umbrella.rss.feed.ArticleActivity">
+    tools:context="org.secfirst.umbrella.rss.views.ArticleActivity">
 
     <item
         android:id="@+id/action_card_list"
