+++ app/src/test/java/org/secfirst/umbrella/RSSFeedServiceTest.java
@@ -10,8 +10,8 @@ import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.shadows.ShadowApplication;
 import org.robolectric.shadows.ShadowLooper;
-import org.secfirst.umbrella.rss.RSSFeedService;
-import org.secfirst.umbrella.rss.feed.CustomFeed;
+import org.secfirst.umbrella.rss.services.RSSFeedService;
+import org.secfirst.umbrella.rss.entities.CustomFeed;
 
 import java.util.List;
